+++ app/models/activity.rb
@@ -1,4 +1,4 @@
 class Activity < ApplicationRecord
   belongs_to :user
-  has_one_attached :fit_file
+  has_one_attached :activity_file
 end